longitude and latitude changed to double

+++ Server/src/main/java/com/begin/dto/ReportDTO.java
@@ -7,28 +7,28 @@ import java.util.Objects;
 
 public class ReportDTO {
 
-  private long latitude;
+  private double latitude;
 
-  private long longitude;
+  private double longitude;
 
   private ZonedDateTime timestamp;
 
   // optional, only if we have incident at the current position
   private List<Value> incidentValues;
 
-  public long getLatitude() {
+  public double getLatitude() {
     return latitude;
   }
 
-  public void setLatitude(long latitude) {
+  public void setLatitude(double latitude) {
     this.latitude = latitude;
   }
 
-  public long getLongitude() {
+  public double getLongitude() {
     return longitude;
   }
 
-  public void setLongitude(long longitude) {
+  public void setLongitude(double longitude) {
     this.longitude = longitude;
   }
